% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appris.R
\docType{class}
\name{APPRIS}
\alias{APPRIS}
\title{APPRIS class}
\format{An \code{\link{R6Class}} generator object}
\usage{
APPRIS
}
\description{
APPRIS class
}
\examples{
appris = APPRIS$new("/Users/jmrodriguez/projects/APPRIS/data/homo_sapiens/ens81.v15/")
appris_m28a3$gene_decision()
}
\keyword{isoforms}
\keyword{principal}

